build(package depencies): updating nextjs and other package dependencies to use the latest versions
All package dependencies have been updated in the package.json to pull down the latest versions.
Fixed a number of linting issues related to eslint default updates.  Added empty getInitialProps to
the preview and visualisation pages as pages now throw error if they are absent - preview and
visualisation are not behaving as expected so they are still WIP.  Fixed the node and npm versions
to 12.0.0 and 6.0.0, respectively